+++ mrv2/lib/mrvGL/mrvGLViewport.h
@@ -90,6 +90,8 @@ namespace mrv
 
         void _drawHelpText() const noexcept;
 
+        void _drawBackground() const noexcept;
+
         void _drawRectangleOutline(
             const math::Box2i& box, const image::Color4f& color,
             const math::Matrix4x4f& mvp) const noexcept;